+++ lib/features/password_manager/forms/note_form/screens/note_form_screen.dart
@@ -11,7 +11,7 @@ import 'package:hoplixi/features/password_manager/dashboard/dashboard.dart';
 import 'package:hoplixi/features/password_manager/dashboard/providers/dashboard_list_refresh_trigger_provider.dart';
 import 'package:hoplixi/features/password_manager/forms/note_form/models/note_form_state.dart';
 import 'package:hoplixi/features/password_manager/forms/note_form/widgets/expandable_quill_modal.dart';
-import 'package:hoplixi/features/password_manager/pickers/vault_item_picker/vault_item_picker_modal.dart';
+import 'package:hoplixi/features/password_manager/pickers/vault_item_picker/vault_item_picker.dart';
 import 'package:hoplixi/shared/ui/button.dart';
 import 'package:hoplixi/shared/ui/modal_sheet_close_button.dart';
 import 'package:hoplixi/shared/utils/vault_link_utils.dart';